+++ optimizations/JPEG/Images/Pixel.cs
@@ -10,7 +10,7 @@ public struct Pixel
 
     private static readonly PixelFormat[] formats = { PixelFormat.RGB, PixelFormat.YCbCr };
 
-    public Pixel(int firstComponent, int secondComponent, int thirdComponent, PixelFormat pixelFormat)
+    public Pixel(short firstComponent, short secondComponent, short thirdComponent, PixelFormat pixelFormat)
     {
         if (!formats.Contains(pixelFormat))
             throw new FormatException("Unknown pixel format: " + pixelFormat);
@@ -31,22 +31,22 @@ public struct Pixel
         }
     }
 
-    private readonly int r;
-    private readonly int g;
-    private readonly int b;
+    private readonly short r;
+    private readonly short g;
+    private readonly short b;
 
-    private readonly int y;
-    private readonly int cb;
-    private readonly int cr;
+    private readonly short y;
+    private readonly short cb;
+    private readonly short cr;
 
-    public int R => format == PixelFormat.RGB ? r : (int)((298.082 * y + 408.583 * Cr) / 256.0 - 222.921);
+    public short R => format == PixelFormat.RGB ? r : (short)((298.082 * y + 408.583 * Cr) / 256.0 - 222.921);
 
-    public int G =>
-        format == PixelFormat.RGB ? g : (int)((298.082 * Y - 100.291 * Cb - 208.120 * Cr) / 256.0 + 135.576);
+    public short G =>
+        format == PixelFormat.RGB ? g : (short)((298.082 * Y - 100.291 * Cb - 208.120 * Cr) / 256.0 + 135.576);
 
-    public int B => format == PixelFormat.RGB ? b : (int)((298.082 * Y + 516.412 * Cb) / 256.0 - 276.836);
+    public short B => format == PixelFormat.RGB ? b : (short)((298.082 * Y + 516.412 * Cb) / 256.0 - 276.836);
 
-    public int Y => format == PixelFormat.YCbCr ? y : (int)(16.0 + (65.738 * R + 129.057 * G + 24.064 * B) / 256.0);
-    public int Cb => format == PixelFormat.YCbCr ? cb : (int)(128.0 + (-37.945 * R - 74.494 * G + 112.439 * B) / 256.0);
-    public int Cr => format == PixelFormat.YCbCr ? cr : (int)(128.0 + (112.439 * R - 94.154 * G - 18.285 * B) / 256.0);
+    public short Y => format == PixelFormat.YCbCr ? y : (short)(16.0 + (65.738 * R + 129.057 * G + 24.064 * B) / 256.0);
+    public short Cb => format == PixelFormat.YCbCr ? cb : (short)(128.0 + (-37.945 * R - 74.494 * G + 112.439 * B) / 256.0);
+    public short Cr => format == PixelFormat.YCbCr ? cr : (short)(128.0 + (112.439 * R - 94.154 * G - 18.285 * B) / 256.0);
 }
